+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 4.1.0 (Unreleased)
 * [Fix autocompletion not working immediately after a dot or the splat operator](https://gitlab.com/Serenata/Serenata/issues/187)
+* [Fix notice "Undefined property `$name` in src/Analysis/VariableScanningVisitor.php on line 117"](https://gitlab.com/Serenata/Serenata/issues/208)
 * [Autocompletion doesn't add imports when function body is missing (but parantheses must be present)](https://gitlab.com/Serenata/Serenata/issues/204)
 * [Fix error "Call to undefined method PhpParser\Node\Stmt\Trait_::isAnonymous()" in traits and interfaces](https://gitlab.com/Serenata/Serenata/issues/206)
 * [Automatically restart without Xdebug enabled if it is present instead of just warning that performance will be degraded](https://gitlab.com/Serenata/Serenata/issues/209)